backend/meropanditlama_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==12.0.0
python-decouple==3.8
django-cors-headers==4.3.0
django-filter==23.3
django-allauth==0.57.0
psycopg2-binary==2.9.10
django-jazzmin==2.6.0